Struct gfx_app::shade::Source

pub struct Source<'a> {
    pub glsl_120: &'a [u8],
    pub glsl_130: &'a [u8],
    pub glsl_140: &'a [u8],
    pub glsl_150: &'a [u8],
    pub glsl_430: &'a [u8],
    pub glsl_es_100: &'a [u8],
    pub glsl_es_200: &'a [u8],
    pub glsl_es_300: &'a [u8],
    pub hlsl_30: &'a [u8],
    pub hlsl_40: &'a [u8],
    pub hlsl_41: &'a [u8],
    pub hlsl_50: &'a [u8],
    pub msl_10: &'a [u8],
    pub msl_11: &'a [u8],
    pub vulkan: &'a [u8],
}

A type storing shader source for different graphics APIs and versions.

Fields
glsl_120: &'a [u8] glsl_130: &'a [u8] glsl_140: &'a [u8] glsl_150: &'a [u8] glsl_430: &'a [u8] glsl_es_100: &'a [u8] glsl_es_200: &'a [u8] glsl_es_300: &'a [u8] hlsl_30: &'a [u8] hlsl_40: &'a [u8] hlsl_41: &'a [u8] hlsl_50: &'a [u8] msl_10: &'a [u8] msl_11: &'a [u8] vulkan: &'a [u8]
Methods

impl<'a> Source<'a>

fn empty() -> Source<'a>

Create an empty shader source. Useful for specifying the remaining structure members upon construction.

fn select(&self, backend: Backend) -> Result<&'a [u8], SelectError>

Pick one of the stored versions that is the highest supported by the backend.